// AIS Project Management Tool/apps/tracker/src/generated/models/aisproject-manager-model.ts
export const AISProjectManagerProjecttypeKeyToLabel = {
  'ProjecttypeKey0': 'ASSET',
  'ProjecttypeKey1': 'DST'
} as const;
export type AISProjectManagerProjecttypeKey = keyof typeof AISProjectManagerProjecttypeKeyToLabel;

export const AISProjectManagerStatusKeyToLabel = {
  'StatusKey0': 'Ideation',
  'StatusKey1': 'In Progress',
  'StatusKey2': 'Completed',
  'StatusKey3': 'De-Prioritised',
  'StatusKey4': 'O & S'
} as const;
export type AISProjectManagerStatusKey = keyof typeof AISProjectManagerStatusKeyToLabel;

export interface AISProjectManager {
  /**
   * @displayName AIS Project Manager
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Project Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  projectname: string;
  /**
   * @displayName Contributors JSON Data
   */
  contributorsjsondata?: string;
  /**
   * @displayName Due Date
   */
  duedate?: string;
  /**
   * @displayName Estimated Man-Hours Saved
   */
  estimatedmanhourssaved?: number | null;
  /**
   * @displayName Expected Benefits
   */
  expectedbenefits?: string;
  /**
   * @displayName Milestones JSON Data
   */
  milestonesjsondata?: string;
  /**
   * @displayName Problem Statement
   */
  problemstatement?: string;
  /**
   * @displayName Project Type
   * @validationRule Required for create/update operations
   */
  projecttypeKey: AISProjectManagerProjecttypeKey;
  /**
   * @displayName Proposed Solution
   */
  proposedsolution?: string;
  /**
   * @displayName Start Date
   */
  startdate?: string;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: AISProjectManagerStatusKey;
}

export const _AISProjectManager = 'AISProjectManager' as const;